refactor: CardsetDetailContent 컴포넌트 분리
카드셋 상세의 핵심 콘텐츠(썸네일, 메타정보, 좋아요/북마크, 관리자 버튼, 학습 설정)를
CardsetDetailContent로 추출. cardset-detail.tsx는 BaseLayout + 소속 그룹 배너만 유지.
applicationRequired 리다이렉트 useEffect 제거 — 시트 방식에서는 리다이렉트 없이 시트를 유지.
